tinycompress: do not loop in the compress read
Return the compress read immediately when it is
successful. client can recall for the remaining bytes.

Also, In case of capture, move the plugin state
to prepared when set params is successful.
Earlier in case of playback this plug state is moved
to prepare only when first write is successful.

CRs-Fixed: 3164253

Change-Id: I2b8e2a12397682c5a17473ea06623c1152cc9819